Create seperate files for medium feature
